% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{\%x=\%}
\alias{\%x=\%}
\alias{\%x=|\%}
\title{Creates string decorators by repeating a pattern}
\usage{
txt \%x=\% n

txt \%x=|\% length.out
}
\arguments{
\item{txt}{Pattern to repeat}

\item{n}{Number of times to repeat the pattern}

\item{length.out}{number of character the output should be}
}
\value{
A character string
}
\description{
Creates string decorators by repeating a pattern
either a given number of times or so that it takes
a given number of character
}
\examples{
"=" \%x=\% 80
"<-+->" \%x=|\% 80

}
