% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_operator.R
\name{f_operator}
\alias{f_operator}
\title{f_operator}
\usage{
f_operator(f)
}
\arguments{
\item{f}{a operator function}
}
\value{
a error message if x is invalid input for operator f,
the result of \emph{f(x)} otherwise
}
\description{
a function operator that transform \emph{sqrt.2()} and \emph{log.2()} into functions such that:
\enumerate{
\item If input is greater than 0, then return normal output.
\item If input is invalid, return an error condition object, with "invalid_input" subclass and invalid value attached.
}
}
\examples{
newSqrt = f_operator(mySqrt)
newLog = f_operator(myLog)
newLog(-1)
newSqrt(-4)
(y = lapply(runif(2,-10,10), newLog))
(y = lapply(runif(2,-10,10), newSqrt))

}
