% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{\%without\%}
\alias{\%without\%}
\title{Remove certain elements from a vector}
\usage{
x \%without\% table
}
\arguments{
\item{x}{Vector}

\item{table}{Elements to remove from \code{x}}
}
\value{
\code{x} without the elements of \code{table}
}
\description{
Remove the elements in \code{table} from \code{s}
}
\examples{
letters \%without\% "a"
}
