% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqrt.R
\name{mySqrt}
\alias{mySqrt}
\title{mySqrt}
\usage{
mySqrt(x)
}
\arguments{
\item{x}{a real number}
}
\value{
a error message if x is negative; the result of square root of x otherwise
}
\description{
mySqrt computes the (principal) square root of non-negative x,
show error message for negative x.
}
\examples{
\dontrun{
a = mySqrt(-1)
b = mySqrt(4)
}

}
