% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{notIn}
\alias{notIn}
\alias{\%!in\%}
\title{not in}
\usage{
x \%!in\% table
}
\arguments{
\item{x}{The values to be matched}

\item{table}{The values to \emph{not} be matched against}
}
\value{
Logical vector, negation of the \code{\%in\%} operators on the same arguments.
}
\description{
Negation of the \code{\%in\%} operator.
}
\examples{
1:10 \%!in\% c(1,3,5,9)

}
