% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{\%of\%}
\alias{\%of\%}
\title{Is an object of a given class}
\usage{
x \%of\% y
}
\arguments{
\item{x}{R object}

\item{y}{Character string, the class to check against.}
}
\value{
Logical value indicating the result
}
\description{
Operator to check if an object is of a given class
}
\examples{
iris \%of\% "data.frame"

}
