\name{operators-package}
\alias{operators-package}
\docType{package}
\title{
Additional binary operators
}
\description{
Operators for common tasks (e.g. regex manipulation) and utilities for working with operators.
}
\details{
  \tabular{ll}{
    Package: \tab operators\cr
    Type: \tab Package\cr
    Version: \tab 0.1-7\cr
    Date: \tab 2010-06-10\cr
    License: \tab GPL-3\cr
  }
}

\author{
  Romain Francois <romain@r-enthusiasts.com>
  Christopher Brown <cbrown@opendatagroup.com>

  Maintainer: Romain Francois and Christopher Brown <operators-authors@r-enthusiasts.com>
}

\section{Notes}{ 
  Submissions welcome.
}

\keyword{ package }

