% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe.R
\name{pipe}
\alias{pipe}
\alias{\%|\%}
\title{Pipe an R object to a unix command}
\usage{
r \%|\% u
}
\arguments{
\item{r}{Any R object}

\item{u}{character string representing the unix command}
}
\value{
An object of S3-class \code{unixoutput}. The \code{print} method
 for \code{unixoutput} objects simply \code{\link{cat}} the
 string.
}
\description{
The operator \code{\link{print}}s the R object into a temporay file and
 then executes the unix command though a \code{\link{pipe}}
}
\examples{
\dontrun{
   rnorm(30) \%|\% 'head -n2'
   rnorm(30) \%|\% 'sed "s/^ *\\\\\\\\[[0-9]*\\\\\\\\]//g" '
}

}
