% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plusEqual.R
\name{plusEqual}
\alias{plusEqual}
\alias{\%+=\%}
\title{Plus Equal Operator}
\usage{
object \%+=\% value
}
\arguments{
\item{object}{object to which to add something}

\item{value}{object to add}
}
\value{
NULL. Used for the side effect of changing the value of \code{object}
}
\description{
Plus equal operator
}
\examples{
x <- 4
x \%+=\% 4
x

}
