\name{operators-package}
\alias{operators-package}
\docType{package}
\title{
Additional binary operators
}
\description{
Additional binary operators and utilities for R
}
\details{
\tabular{ll}{
Package: \tab operators\cr
Type: \tab Package\cr
Version: \tab 0.1-7\cr
Date: \tab 2010-06-07\cr
License: \tab GPL-3\cr
}
}
\author{
Romain Francois <romain@r-enthusiasts.com>
Christopher Brown <cbrown@opendatagroup.com>

Maintainer: Romain Francois and Christopher Brown <operators-authors@r-enthusiasts.com>
}
\keyword{ package }

