\name{operators-package}
\alias{operators-package}
\alias{operators}
\docType{package}
\title{
Additional binary operators
}
\description{
Additional binary operators to expand R's. These are basically 
convenience functions.
}
\details{
\tabular{ll}{
Package: \tab operators\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2007-06-05\cr
License: \tab GPL3\cr
}
}
\author{
Romain Francois <rfrancois@mango-solutions.com>

Maintainer: Romain Francois <rfrancois@mango-solutions.com>
}
\keyword{ package }

