\name{operators-package}
\alias{operators-package}
\alias{operators}
\docType{package}
\title{
Additional binary operators
}
\description{
Additional binary operators for R
}
\details{
\tabular{ll}{
Package: \tab operators\cr
Type: \tab Package\cr
Version: \tab 0.0-16\cr
Date: \tab 2009-01-22\cr
License: \tab GPL3\cr
}
}
\author{
Romain Francois <francoisromain@free.fr>

Maintainer: Romain Francois <francoisromain@free.fr>
}
\keyword{ package }

