\name{operators-package}
\alias{operators-package}
\alias{operators}
\docType{package}
\title{
Additional binary operators
}
\description{
Additional binary operators for R
}
\author{
Romain Francois <romain@r-enthusiasts.com>

Maintainer: Romain Francois <romain@r-enthusiasts.com>
}
\keyword{ package }

