\name{operators-package}
\alias{operators-package}
\alias{operators}
\docType{package}
\title{
Additional binary operators
}
\description{
Additional binary operators for R
}
\details{
\tabular{ll}{
Package: \tab operators\cr
Type: \tab Package\cr
Version: \tab 0.1-6\cr
Date: \tab 2010-04-11\cr
License: \tab GPL-3\cr
}
}
\author{
Romain Francois <romain@r-enthusiasts.com>

Maintainer: Romain Francois <romain@r-enthusiasts.com>
}
\keyword{ package }

